# SinkBuffer optimization by removing waiting logic of rpc task