--- a/be/src/exec/pipeline/exchange/sink_buffer.h
+++ b/be/src/exec/pipeline/exchange/sink_buffer.h
@@ -16,6 +16,7 @@
 #include "util/brpc_stub_cache.h"
 #include "util/callback_closure.h"
 #include "util/defer_op.h"
+#include "util/phmap/phmap.h"
 #include "util/priority_thread_pool.hpp"
 
 namespace starrocks::pipeline {
@@ -41,14 +42,14 @@ public:
         for (const auto& dest : destinations) {
             const auto& dest_instance_id = dest.fragment_instance_id;
 
-            auto it = _num_sinkers_per_dest_instance.find(dest_instance_id);
+            auto it = _num_sinkers_per_dest_instance.find(dest_instance_id.lo);
             if (it != _num_sinkers_per_dest_instance.end()) {
                 it->second += num_sinkers;
             } else {
-                _num_sinkers_per_dest_instance[dest_instance_id] = num_sinkers;
+                _num_sinkers_per_dest_instance[dest_instance_id.lo] = num_sinkers;
 
                 // This dest_instance_id first occurs, so create other variable for this dest instance.
-                auto* closure = new CallBackClosure<PTransmitChunkResult>();
+                auto* closure = new CallBackClosure<PTransmitChunkResult, TUniqueId>();
                 closure->ref();
                 closure->addFailedHandler([this]() noexcept {
                     _in_flight_rpc_num.fetch_sub(1, std::memory_order_release);
@@ -58,7 +59,7 @@ public:
                     }
                     LOG(WARNING) << " transmit chunk rpc failed";
                 });
-                closure->addSuccessHandler([this](const PTransmitChunkResult& result) noexcept {
+                closure->addSuccessHandler([this, closure](const PTransmitChunkResult& result) noexcept {
                     _in_flight_rpc_num.fetch_sub(1, std::memory_order_release);
                     Status status(result.status());
                     if (!status.ok()) {
@@ -67,17 +68,22 @@ public:
                             _is_finishing = true;
                         }
                         LOG(WARNING) << " transmit chunk rpc failed, " << status.message();
+                    } else {
+                        _is_closure_finishing[closure->context().lo]->store(true, std::memory_order_release);
+                        _try_to_trigger_rpc_task(closure->context());
                     }
                 });
-                _closures[dest_instance_id] = closure;
+                _closures[dest_instance_id.lo] = closure;
+                _is_closure_finishing[dest_instance_id.lo] = std::make_unique<std::atomic_bool>(false);
 
-                _buffers[dest_instance_id] = std::queue<TransmitChunkInfo, std::list<TransmitChunkInfo>>();
-                _request_seqs[dest_instance_id] = 0;
+                _request_seqs[dest_instance_id.lo] = 0;
+                _buffers[dest_instance_id.lo] = std::queue<TransmitChunkInfo, std::list<TransmitChunkInfo>>();
+                _instance_mutexes[dest_instance_id.lo] = std::make_unique<std::mutex>();
 
                 PUniqueId finst_id;
                 finst_id.set_hi(dest_instance_id.hi);
                 finst_id.set_lo(dest_instance_id.lo);
-                _instance_id2finst_id[dest_instance_id] = std::move(finst_id);
+                _instance_id2finst_id[dest_instance_id.lo] = std::move(finst_id);
             }
 
             _expected_eos = 0;
@@ -88,6 +94,8 @@ public:
     }
 
     ~SinkBuffer() {
+        _is_destructing.store(true, std::memory_order_release);
+
         // Handle abnormal exit of SinkBuffer
         // TODO(hcf) It may be solved by adding state 'PENDING_FINISH' for ExchangeSinkOperator
         bool abnormal_exit = false;
@@ -98,8 +106,7 @@ public:
         }
         if (abnormal_exit) {
             for (auto& [_, closure] : _closures) {
-                auto cntl = &closure->cntl;
-                brpc::Join(cntl->call_id());
+                brpc::Join(closure->cntl.call_id());
             }
         }
 
@@ -122,8 +129,8 @@ public:
         }
         for (auto& [_, buffer] : _buffers) {
             while (!buffer.empty()) {
-                auto& info = buffer.front();
-                info.params->release_finst_id();
+                auto& request = buffer.front();
+                request.params->release_finst_id();
                 buffer.pop();
             }
         }
@@ -131,18 +138,16 @@ public:
 
     void add_request(const TransmitChunkInfo& request) {
         DCHECK(_expected_eos > 0);
+        if (_is_destructing.load(std::memory_order_acquire)) {
+            request.params->release_finst_id();
+            return;
+        }
         {
-            // Guarantee that no request will be added to buffer after finishing stage
-            std::lock_guard<std::mutex> l(_mutex);
-            if (_is_finishing) {
-                request.params->release_finst_id();
-                return;
-            } else {
-                _buffers[request.fragment_instance_id].push(request);
-            }
+            std::lock_guard<std::mutex> l(*_instance_mutexes[request.fragment_instance_id.lo]);
+            _buffers[request.fragment_instance_id.lo].push(request);
         }
 
-        _try_to_trigger_rpc_task();
+        _try_to_trigger_rpc_task(request.fragment_instance_id);
     }
 
     bool is_full() const {
@@ -165,16 +170,15 @@ public:
     }
 
 private:
-    void _try_to_trigger_rpc_task() {
+    void _try_to_trigger_rpc_task(const TUniqueId& instance_id) {
         // Guarantee that no new rpc task will be created after finishing stage
         std::lock_guard<std::mutex> l(_mutex);
         if (_is_finishing) {
             return;
         }
 
-        // Tell rpc task scan buffer one more time
+        _events.push(instance_id);
         if (_is_rpc_task_active) {
-            _need_one_more_check = true;
             return;
         }
 
@@ -200,49 +204,45 @@ private:
             SCOPED_THREAD_LOCAL_MEM_TRACKER_SETTER(_mem_tracker);
 
             for (;;) {
-                bool buffer_empty = true;
-                bool send_any = false;
-                for (auto& [fragment_instance_id, buffer] : _buffers) {
-                    auto* closure = _closures[fragment_instance_id];
-                    DCHECK(closure != nullptr);
-                    if (buffer.empty()) {
-                        continue;
+                TUniqueId instance_id;
+                {
+                    std::lock_guard<std::mutex> l(_mutex);
+                    if (_is_finishing || _events.empty()) {
+                        _is_rpc_task_active = false;
+                        break;
                     }
-                    buffer_empty = false;
-                    if (closure->has_in_flight_rpc()) {
-                        continue;
-                    }
-                    send_any = true;
-
-                    TransmitChunkInfo info = buffer.front();
-                    _send_rpc(info);
-                    info.params->release_finst_id();
-                    {
-                        std::lock_guard<std::mutex> l(_mutex);
-                        buffer.pop();
-                    }
+                    instance_id = _events.front();
+                    _events.pop();
                 }
 
-                if (buffer_empty) {
-                    // When traversing the buffer, a new request may be inserted
-                    // so the value of buffer_empty may not be that accurate, we
-                    // need to communicate with add_request through _need_one_more_check
-                    // in order to avoid exiting rpc task with non-empty buffer
-                    std::lock_guard<std::mutex> l(_mutex);
-                    if (_need_one_more_check) {
-                        _need_one_more_check = false;
-                        continue;
+                auto* closure = _closures[instance_id.lo];
+                auto& buffer = _buffers[instance_id.lo];
+                {
+                    for (;;) {
+                        {
+                            std::lock_guard<std::mutex> l(*_instance_mutexes[instance_id.lo]);
+                            if (buffer.empty()) {
+                                break;
+                            }
+                        }
+                        if (closure->has_in_flight_rpc()) {
+                            // This closure is bound to finish for a short while
+                            // So block wait for the closure to be finished
+                            if (_is_closure_finishing[instance_id.lo]->load(std::memory_order_acquire)) {
+                                brpc::Join(closure->cntl.call_id());
+                                _is_closure_finishing[instance_id.lo]->store(false, std::memory_order_release);
+                            } else {
+                                break;
+                            }
+                        }
+                        TransmitChunkInfo request = buffer.front();
+                        _send_rpc(request);
+                        request.params->release_finst_id();
+                        {
+                            std::lock_guard<std::mutex> l(*_instance_mutexes[instance_id.lo]);
+                            buffer.pop();
+                        }
                     }
-                    _need_one_more_check = false;
-                    _is_rpc_task_active = false;
-                    break;
-                } else if (!send_any) {
-#ifdef __x86_64__
-                    _mm_pause();
-#else
-                    // TODO: Maybe there's a better intrinsic like _mm_pause on non-x86_64 architecture.
-                    sched_yield();
-#endif
                 }
             }
         } catch (const std::exception& exp) {
@@ -267,7 +267,7 @@ private:
             // Only the last eos is sent to ExchangeSourceOperator. it must be guaranteed that
             // eos is the last packet to send to finish the input stream of the corresponding of
             // ExchangeSourceOperator and eos is sent exactly-once.
-            if (--_num_sinkers_per_dest_instance[request.fragment_instance_id] > 0) {
+            if (--_num_sinkers_per_dest_instance[request.fragment_instance_id.lo] > 0) {
                 if (request.params->chunks_size() == 0) {
                     return;
                 } else {
@@ -276,16 +276,17 @@ private:
             }
         }
 
-        request.params->set_allocated_finst_id(&_instance_id2finst_id[request.fragment_instance_id]);
-        request.params->set_sequence(_request_seqs[request.fragment_instance_id]++);
+        request.params->set_allocated_finst_id(&_instance_id2finst_id[request.fragment_instance_id.lo]);
+        request.params->set_sequence(_request_seqs[request.fragment_instance_id.lo]++);
 
-        auto* closure = _closures[request.fragment_instance_id];
+        auto* closure = _closures[request.fragment_instance_id.lo];
         DCHECK(closure != nullptr);
         DCHECK(!closure->has_in_flight_rpc());
         closure->ref();
         closure->cntl.Reset();
         closure->cntl.set_timeout_ms(_brpc_timeout_ms);
         closure->cntl.request_attachment().append(request.attachment);
+        closure->set_context(request.fragment_instance_id);
         _in_flight_rpc_num.fetch_add(1, std::memory_order_release);
         request.brpc_stub->transmit_chunk(&closure->cntl, request.params.get(), &closure->result, closure);
     }
@@ -294,22 +295,36 @@ private:
     MemTracker* _mem_tracker = nullptr;
     const int32_t _brpc_timeout_ms;
 
-    std::unordered_map<TUniqueId, size_t> _num_sinkers_per_dest_instance;
-    std::unordered_map<TUniqueId, int64_t> _request_seqs;
+    // Taking into account of efficiency, all the following maps
+    // use int64_t as key, which is the field type of TUniqueId::lo
+    // because TUniqueId::hi is exactly the same in one query
+
+    phmap::flat_hash_map<int64_t, size_t> _num_sinkers_per_dest_instance;
+    phmap::flat_hash_map<int64_t, int64_t> _request_seqs;
     std::atomic<int32_t> _in_flight_rpc_num = 0;
 
     // The request needs the reference to the allocated finst id,
     // so cache finst id for each dest fragment instance.
-    std::unordered_map<TUniqueId, PUniqueId> _instance_id2finst_id;
+    phmap::flat_hash_map<int64_t, PUniqueId> _instance_id2finst_id;
 
-    std::unordered_map<TUniqueId, CallBackClosure<PTransmitChunkResult>*> _closures;
-    std::unordered_map<TUniqueId, std::queue<TransmitChunkInfo, std::list<TransmitChunkInfo>>> _buffers;
+    phmap::flat_hash_map<int64_t, CallBackClosure<PTransmitChunkResult, TUniqueId>*> _closures;
+    // Closure becomes idle only when handle finished execution
+    // Rpc task may be created in success handler of closure, if this task being executed immediately
+    // and judge whether closure is idle by method closure->has_in_flight_rpc(), and this method may
+    // return true if success handler hasn't finished its execution, which lead to this event(channel ready)
+    // being swallowed
+    phmap::flat_hash_map<int64_t, std::unique_ptr<std::atomic_bool>> _is_closure_finishing;
+    phmap::flat_hash_map<int64_t, std::queue<TransmitChunkInfo, std::list<TransmitChunkInfo>>> _buffers;
+    phmap::flat_hash_map<int64_t, std::unique_ptr<std::mutex>> _instance_mutexes;
 
     std::mutex _mutex;
+    // Events include 'new request' and 'channel ready'
+    std::queue<TUniqueId, std::list<TUniqueId>> _events;
     bool _is_finishing = false;
-    bool _need_one_more_check = false;
     bool _is_rpc_task_active = false;
     int32_t _expected_eos = 0;
+
+    std::atomic_bool _is_destructing = false;
 
     // This field is used to help properly handle special situation
     // that SinkBuffer may destruct when method is_finished() return false