**Before Optomize**

* At any moment, each SinkBuffer has at most one active rpc task, and this rpc task will not exit if buffer is not empty, which means that if closure is busy, the rpc task will wait for the closure to be finished.
* In high concurrency scenarios, these rpc task will monopolize the execution thread, which may cause other SinkBuffers to starve

**After Optimize**

* Introduce a queue of events, include two type of event
    1. new request arriving
    2. channel becomes ready
* Allow each SinkBuffer still has at most one active rpc task, but this rpc task will exit if event queue is empty
* Reduce exchange thread pool size from hard core size to 4, which means less threads can support the same scale of concurrency

**Test Info**

* 1 fe and 3 be
    * 64c/128g
    * fe mixed with one of be
* ssb test set
    * scale factor: 100
* test sqls and command are as follows

```sh
mysqlslap -u root -psr -h 1.1.1.1 -P 9999 -c 64 --number-of-queries 128 -i 10 --create-schema=ssb --query="<sql>";
```

```sql
-- Q1
SELECT SUM(lo_quantity) FROM lineorder_flat;
```

**Test Result**

---

| Q1 | -c 64 --number-of-queries 128 -i 10 | -c 128 --number-of-queries 256 -i 10 |
|:--|:--|:--|
| pipeline_exchange_thread_pool_thread_num=1 | before: 3.005s</br>after: 2.927s</br> rate: 102.7% | before: 6.168s</br>after: 6.010s</br> rate: 102.6% |
| pipeline_exchange_thread_pool_thread_num=2 | before: 2.848s</br>after: 2.796s</br> rate: 101.9% | before: 5.851s</br>after: 5.876s</br> rate: 99.6% |
| pipeline_exchange_thread_pool_thread_num=4 | before: 2.800s</br>after: 2.797s</br> rate: 100.1% | before: 5.904s</br>after: 5.882s</br> rate: 100.4% |
| pipeline_exchange_thread_pool_thread_num=8 | before: 2.914s</br>after: 2.893s</br> rate: 100.7% | before: 5.997s</br>after: 5.978s</br> rate: 100.3% |
| pipeline_exchange_thread_pool_thread_num=16 | before: 2.954s</br>after: 2.891s</br> rate: 102.2% | before: 6.075s</br>after: 5.955s</br> rate: 102.0% |
| pipeline_exchange_thread_pool_thread_num=32 | before: 2.987s</br>after: 2.886s</br> rate: 103.5% | before: 6.242s</br>after: 6.023s</br> rate: 103.6% |
| pipeline_exchange_thread_pool_thread_num=64 | before: 2.982s</br>after: 2.860s</br> rate: 104.3% | before: 6.189s</br>after: 5.914s</br> rate: 104.6% |


**Conclusion**

1. The optimization effect is not that significant, it can improve the performance of about 1-5% in the case of concurrency
2. As the number of thread pools increases, performance remains basically unchanged at the current scale of concurrency. So the recommended thread num can be set to 2 considering the resource utilization efficiency